Structure fix
