+++ src/components/atoms/Button/index.js
@@ -4,15 +4,8 @@ import {View, Text, StyleSheet, TouchableOpacity} from 'react-native';
 const Button = props => {
   const {onPress, text} = props;
   return (
-    <TouchableOpacity
-      style = {styles.buttonContainer}
-      onPress = {onPress}
-    >
-      <Text 
-      style={styles.buttonText}
-      >
-        { text }
-      </Text>
+    <TouchableOpacity style={styles.buttonContainer} onPress={onPress}>
+      <Text style={styles.buttonText}>{text}</Text>
     </TouchableOpacity>
   );
 };
