+++ src/screens/Welcome.js
@@ -1,22 +1,19 @@
 import React from 'react';
 import {SafeAreaView, Text, StyleSheet, Image} from 'react-native';
-import Icon from 'react-native-vector-icons/FontAwesome';
-import Button from '../components/Button';
+import Button from '../components/atoms/Button';
+import LogoImg from '../assets/images/LogoUDC2.png';
 
-export const Welcome = ({ navigation }) => {
+export const Welcome = ({navigation}) => {
   return (
     <SafeAreaView style={styles.container}>
       <Text style={styles.title}>Guardaditos</Text>
       <Text>By TimRocket</Text>
-      <Image
-          style={styles.tinyLogo}
-          source={{
-          uri: 'https://cdn-icons-png.flaticon.com/128/3566/3566826.png',
-          }}
-      />
-      <Button 
-      text = "¡Comenzar!" 
-      onPress = { () => { navigation.navigate('Login') }}
+      <Image style={styles.tinyLogo} source={LogoImg} />
+      <Button
+        text="¡Comenzar!"
+        onPress={() => {
+          navigation.navigate('Login');
+        }}
       />
     </SafeAreaView>
   );
@@ -36,9 +33,9 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
   },
   tinyLogo: {
-    width: 60,
-    height: 60,
-    marginTop: 15,
-    marginBottom: 50,
+    width: 100,
+    height: 100,
+    marginTop: 20,
+    marginBottom: 20,
   },
 });
