+++ src/screens/index.js
@@ -1,0 +1,4 @@
+export {Home} from './Home';
+export {Login} from './Login';
+export {Register} from './Register';
+export {Welcome} from './Welcome';
